Metadata of samples from OTU tables (N-ICE, 2015)

This dataset provides information on the samples used for the N-ICE dataset.

Usage

nice_env

Format

nice_tidy

A data frame with 4716 rows and 10 columns:

Sample

Sample ID used in the original study

ENA_ID

Sample ID equivalent to Sample in the OTU table

Month

Month of sampling

Region

Ocean region of sampling event

Water.mass

Water mass of sampling event

Latitude

Latitude of sampling event

Longitude

Longitude of sampling event

Source

https://link.springer.com/article/10.1007/s00248-021-01821-9

Details

Based on de Sousa et al., 2019.

References

de Sousa, A. G. G., Tomasino, M. P., Duarte, P., Fernández-Méndez, M., Assmy, P., Ribeiro, H., Surkont, J., Leite, R. B., Pereira-Leal, J. B., Torgo, L., & Magalhães, C. (2019). Diversity and Composition of Pelagic Prokaryotic and Protist Communities in a Thin Arctic Sea-Ice Regime. Microbial Ecology, 78(2), 388–408.

See also